feat(bootstrap): Phase 6 — wiring simplification via DI containers
- Extract bundle dataclasses (VoiceBundle, BotAuthBundle, CliPoolBundle,
  BuildHubDeps, RegisterAgentsDeps, WireAdaptersDeps) to types.py
- Add _build_hub, _init_clipool, _build_hub_and_wire to hub_builder.py
- Add _init_bot_auths_and_agents to agent_factory.py
- Extract standalone Telegram/Discord wiring to wiring/standalone_*.py
- Merge clipool + turn_writer standalones into worker_standalone.py
- Extract lifecycle helpers to lifecycle_helpers.py
- Extract health server to infra/health.py
- Create tests/fakes/ with FakeNatsClient, FakeStore, FakeClaudeCliDriver,
  FakeStt, FakeTts, FakeNkeyProvider, MockAdapter
- Fix all test imports and monkeypatch targets for moved symbols
- Update .importlinter with tests-fakes-isolation rule
- wiring-bootstrap-deps count: 0 (was 54)
