+++ src/lyra/bootstrap/factory/wiring_helpers.py
@@ -6,119 +6,37 @@ from __future__ import annotations
 
 import asyncio
 import logging
-from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from lyra.bootstrap.factory.agent_factory import _resolve_agents, _resolve_bot_agent_map
+from lyra.bootstrap.factory.agent_factory import _resolve_agents
 from lyra.bootstrap.factory.config import (
     MessageIndexConfig,
-    _build_agent_overrides,
-    _load_circuit_config,
-    _load_cli_pool_config,
-    _load_debouncer_config,
-    _load_event_bus_config,
-    _load_hub_config,
     _load_inbound_bus_config,
     _load_llm_config,
-    _load_messages,
     _load_pairing_config,
-    _load_pool_config,
     build_adapter_config_bundle,
 )
-from lyra.bootstrap.types import DiscordAdapterEntry, WiredAdapters
+from lyra.bootstrap.types import (
+    DiscordAdapterEntry,
+    RegisterAgentsDeps,
+    VoiceBundle,
+    WireAdaptersDeps,
+    WiredAdapters,
+)
 from lyra.bootstrap.wiring.bootstrap_wiring import (
-    _build_bot_auths,
     wire_discord_adapters,
     wire_telegram_adapters,
 )
-from lyra.config import load_multibot_config
-from lyra.core.agent import Agent
-from lyra.core.agent.agent_loader import agent_row_to_config
-from lyra.core.cli.cli_pool import CliPool
-from lyra.core.config import HubConfig
-from lyra.core.hub import Hub
-from lyra.core.hub.event_bus import PipelineEventBus
 from lyra.core.messaging.message import InboundMessage
-from lyra.infrastructure.audit import JetStreamAuditSink
-from lyra.infrastructure.resume_publisher_adapter import TurnPublisherAdapter
 from lyra.infrastructure.stores.pairing import PairingManager, set_pairing_manager
 from lyra.nats.nats_bus import NatsBus
 from lyra.nats.queue_groups import HUB_INBOUND
-from lyra.transport.turn_publisher import TurnPublisher
 
 if TYPE_CHECKING:
     import nats
-    from lyra.llm.llm_client import LlmClient
 
 log = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# Return bundles (plain dataclasses — no framework imports)
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class VoiceBundle:
-    stt_service: object
-    tts_service: object
-    nats_llm_client: "LlmClient | None"
-
-
-@dataclass
-class BotAuthBundle:
-    tg_bot_auths: object
-    dc_bot_auths: object
-    bot_agent_map: dict
-    agent_configs: dict[str, Agent]
-    first_agent_config: Agent
-    msg_manager: object
-    circuit_registry: object
-    admin_user_ids: list
-
-
-@dataclass
-class CliPoolBundle:
-    cli_pool: CliPool
-    cli_nats_driver: "LlmClient | None"
-    worker: object
-    audit_sink: JetStreamAuditSink
-
-
-# ---------------------------------------------------------------------------
-# DI containers
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class BuildHubDeps:
-    raw_config: dict
-    bundle: BotAuthBundle
-    voice: VoiceBundle
-    inbound_bus: NatsBus
-    pm: PairingManager | None
-    stores: object
-
-
-@dataclass
-class RegisterAgentsDeps:
-    hub: Hub
-    bundle: BotAuthBundle
-    voice: VoiceBundle
-    clipool: CliPoolBundle
-    raw_config: dict
-    stores: object
-
-
-@dataclass
-class WireAdaptersDeps:
-    hub: Hub
-    bundle: BotAuthBundle
-    nc: nats.aio.client.Client
-    stores: object
-    vault_dir: Path
-    raw_config: dict
 
 
 # ---------------------------------------------------------------------------
@@ -160,83 +78,9 @@ async def _prune_message_index(stores: object, raw_config: dict) -> None:
         )
 
 
-async def _init_bot_auths_and_agents(
-    stores: object,
-    raw_config: dict,
-) -> BotAuthBundle:
-    """Resolve multibot config, build authenticators, load agent configs."""
-    circuit_registry, admin_user_ids = _load_circuit_config(raw_config)
-
-    try:
-        tg_multi_cfg, dc_multi_cfg = load_multibot_config(raw_config)
-    except ValueError as exc:
-        raise SystemExit(str(exc))
-
-    from lyra.bootstrap.wiring.bootstrap_wiring import BotAuthDeps
-
-    tg_bot_auths, dc_bot_auths = _build_bot_auths(
-        BotAuthDeps(
-            bot_store=stores.bot,
-            tg_multi_cfg=tg_multi_cfg,
-            dc_multi_cfg=dc_multi_cfg,
-            auth_store=stores.auth,
-            admin_user_ids=admin_user_ids,
-            alias_store=stores.identity_alias,
-        )
-    )
-    log.info(
-        "Authenticator: %d admin_user_id(s) configured",
-        len(admin_user_ids),
-    )
-
-    if not tg_bot_auths and not dc_bot_auths:
-        raise SystemExit(
-            "No adapters configured — add at least one"
-            " [[telegram.bots]] or [[discord.bots]] entry"
-            " and run 'lyra bot init' to seed the bot store"
-        )
-
-    bot_agent_map = await _resolve_bot_agent_map(
-        stores.agent, tg_multi_cfg.bots, dc_multi_cfg.bots
-    )
-    agent_names: set[str] = set(bot_agent_map.values())
-
-    agent_configs: dict[str, Agent] = {}
-    for n in sorted(agent_names):
-        row = stores.agent.get(n)
-        if row is not None:
-            overrides = _build_agent_overrides(raw_config, n)
-            agent_configs[n] = agent_row_to_config(
-                row,
-                instance_overrides=overrides.model_dump(),
-            )
-        else:
-            log.error("Agent %r not found in DB — skipping", n)
-    if not agent_configs:
-        raise SystemExit(
-            "No agent configs could be loaded — run"
-            " 'lyra agent init' to seed the agents table"
-        )
-
-    first_agent_name = next(iter(sorted(agent_configs)))
-    first_agent_config = agent_configs[first_agent_name]
-    msg_manager = _load_messages(language=first_agent_config.i18n_language)
-
-    return BotAuthBundle(
-        tg_bot_auths=tg_bot_auths,
-        dc_bot_auths=dc_bot_auths,
-        bot_agent_map=bot_agent_map,
-        agent_configs=agent_configs,
-        first_agent_config=first_agent_config,
-        msg_manager=msg_manager,
-        circuit_registry=circuit_registry,
-        admin_user_ids=admin_user_ids,
-    )
-
-
 async def _init_pairing(
     raw_config: dict,
-    admin_user_ids: list,
+    admin_user_ids: frozenset[str],
     vault_dir: Path,
     stores: object,
 ) -> PairingManager | None:
@@ -277,95 +121,6 @@ async def _init_voice_services(
         stt_service=stt_service,
         tts_service=tts_service,
         nats_llm_client=nats_llm_client,
-    )
-
-
-def _build_hub(deps: BuildHubDeps) -> Hub:
-    """Construct HubConfig and Hub, wire stores and alias store."""
-    cli_pool_cfg = _load_cli_pool_config(deps.raw_config)
-    hub_cfg = _load_hub_config(deps.raw_config)
-    pool_cfg = _load_pool_config(deps.raw_config)
-    debouncer_cfg = _load_debouncer_config(deps.raw_config)
-    event_bus_cfg = _load_event_bus_config(deps.raw_config)
-    inbound_bus_cfg = _load_inbound_bus_config(deps.raw_config)
-    event_bus = PipelineEventBus(maxsize=event_bus_cfg.queue_maxsize)
-
-    hub_config = HubConfig(
-        rate_limit=hub_cfg.rate_limit,
-        rate_window=hub_cfg.rate_window,
-        pool_ttl=hub_cfg.pool_ttl,
-        debounce_ms=debouncer_cfg.default_debounce_ms,
-        cancel_on_new_message=debouncer_cfg.cancel_on_new_message,
-        turn_timeout=cli_pool_cfg.turn_timeout,
-        safe_dispatch_timeout=pool_cfg.safe_dispatch_timeout,
-        staging_maxsize=inbound_bus_cfg.staging_maxsize,
-        platform_queue_maxsize=inbound_bus_cfg.platform_queue_maxsize,
-        queue_depth_threshold=inbound_bus_cfg.queue_depth_threshold,
-        max_merged_chars=debouncer_cfg.max_merged_chars,
-    )
-
-    # Wire TurnPublisher + ResumePublisherAdapter before Hub construction
-    js = deps.inbound_bus._nc.jetstream()
-    turn_publisher = TurnPublisher(js)
-    adapter = TurnPublisherAdapter(turn_publisher, deps.stores.turn)
-
-    hub = Hub(
-        circuit_registry=deps.bundle.circuit_registry,
-        msg_manager=deps.bundle.msg_manager,
-        pairing_manager=deps.pm,
-        stt=deps.voice.stt_service,
-        tts=deps.voice.tts_service,
-        prefs_store=deps.stores.prefs,
-        event_bus=event_bus,
-        inbound_bus=deps.inbound_bus,
-        config=hub_config,
-        resume_publisher=adapter,
-    )
-    hub.set_turn_store(deps.stores.turn)
-    hub.set_message_index(deps.stores.message_index)
-
-    # Wire alias_store (#472)
-    deps.stores.prefs.set_alias_store(deps.stores.identity_alias)
-    hub.set_alias_store(deps.stores.identity_alias)
-    hub.set_turn_publisher(turn_publisher)
-
-    return hub
-
-
-async def _init_clipool(
-    nc: nats.aio.client.Client,
-    raw_config: dict,
-    stores: object,
-) -> CliPoolBundle:
-    """Build LlmClient (clipool), CliPool, CliPoolNatsWorker."""
-    from lyra.adapters.clipool.clipool_worker import CliPoolNatsWorker
-    from lyra.bootstrap.factory.hub_builder import build_llm_client
-
-    cli_pool_cfg = _load_cli_pool_config(raw_config)
-    audit_sink = JetStreamAuditSink()
-    await audit_sink.provision(nc)
-
-    cli_nats_driver = await build_llm_client(nc)
-    cli_pool = CliPool(
-        idle_ttl=cli_pool_cfg.idle_ttl,
-        default_timeout=cli_pool_cfg.default_timeout,
-        reaper_interval=cli_pool_cfg.reaper_interval,
-        kill_timeout=cli_pool_cfg.kill_timeout,
-        read_buffer_bytes=cli_pool_cfg.read_buffer_bytes,
-        stdin_drain_timeout=cli_pool_cfg.stdin_drain_timeout,
-        max_idle_retries=cli_pool_cfg.max_idle_retries,
-        intermediate_timeout=cli_pool_cfg.intermediate_timeout,
-        audit_sink=audit_sink,
-    )
-    await cli_pool.start()
-    cli_pool.set_turn_store(stores.turn)
-    worker = CliPoolNatsWorker(cli_pool, timeout=cli_pool_cfg.default_timeout)
-
-    return CliPoolBundle(
-        cli_pool=cli_pool,
-        cli_nats_driver=cli_nats_driver,
-        worker=worker,
-        audit_sink=audit_sink,
     )
