+++ tests/test_bootstrap_missing_credentials.py
@@ -46,11 +46,11 @@ async def test_adapter_raises_bootstrap_error_on_missing_token(
         patch("nats.connect", AsyncMock(return_value=mock_nc)),
         patch("lyra.nats.nats_bus.NatsBus", return_value=mock_inbound_bus),
         patch(
-            "lyra.bootstrap.standalone.adapter_standalone.NatsOutboundListener",
+            "lyra.bootstrap.wiring.standalone_telegram.NatsOutboundListener",
             return_value=AsyncMock(),
         ),
         patch(
-            "lyra.bootstrap.standalone.adapter_standalone.wait_for_hub",
+            "lyra.bootstrap.wiring.standalone_telegram.wait_for_hub",
             AsyncMock(return_value=True),
         ),
         patch("lyra.bootstrap.credentials._is_prod_env", return_value=False),
