+++ tools/file_exemptions.txt
@@ -26,3 +26,5 @@ src/lyra/nats/nats_channel_proxy.py  # 361 lines — DEBT:file-exemptions — #6
 src/lyra/cli.py  # 311 lines — DEBT:file-exemptions — #1331 +12 lines for `lyra turn-writer` subcommand (T28); #1453 +5 lines for importlib subcommand registration (moved from cli_bot/cli_agent module-exit); #1415 +1 line for platform subcommand registration. Linear growth with subcommand count; consider migrating to per-subcommand modules when next subcommand lands.
 src/lyra/core/commands/command_router.py  # 310 lines — DEBT:file-exemptions — #1434 base_dir param + _base_dir wiring for admin path-traversal fix
 src/lyra/monitoring/checks.py  # 306 lines — DEBT:file-exemptions — #1068 +6 lines (blobstore disk_pct + inode_pct checks)
+src/lyra/bootstrap/factory/agent_factory.py  # 355 lines — DEBT:file-exemptions — #1283 Phase 6: _init_bot_auths_and_agents + _resolve_bot_agent_map + _create_agent + _build_provider_registry extracted into agent_factory.py (bootstrap DI container migration)
+src/lyra/bootstrap/factory/hub_builder.py  # 403 lines — DEBT:file-exemptions — #1283 Phase 6: _build_hub + _init_clipool + _build_hub_and_wire extracted into hub_builder.py (bootstrap DI container migration); pre-existing build_hub + build_inbound_bus + build_llm_client
